refactoring: fixed layout
