+++ rs-react-app/src/app/[locale]/404/layout.tsx
@@ -1,0 +1,11 @@
+import '../../../styles/general/App.scss';
+
+import ErrorPage from './page';
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      <ErrorPage></ErrorPage>
+    </>
+  );
+}
