+++ rs-react-app/src/app/[locale]/page.tsx
@@ -1,11 +1,11 @@
 import {getTranslations} from 'next-intl/server';
 
+import ContentBlock from '../../components/contentBlock/contentBlock';
+
 const t = await getTranslations('HomePage')
 
 export default function HomePage() {
   return (
-    <main>
-      <h1>{t('title')}</h1>
-    </main>
+    <ContentBlock></ContentBlock>
   )
 }
